Users and Events fully functional
some changes will be made in the future + added functions.
but the classes are fully functional

+++ DB_event.php
@@ -1,11 +1,15 @@
 <?php
 
-require_once('DB.php');
 require_once('DB_user.php');
+require_once('DB_rsvp.php');
 
 interface iEvent {
 
     public function deleteEvent(User $user);
+
+    public function getEventID();
+
+    public function getDB();
 }
 
 class Event implements iEvent {
@@ -16,83 +20,32 @@ class Event implements iEvent {
     public $messages;
     public $rawData;
 
-    /*
-      private function createRSVP(int $eventID) {
-      $sql = "CREATE TABLE IF NOT EXISTS RSVP_'$eventID' (
-      ID INT(5) NOT NULL AUTO_INCREMENT PRIMARY KEY,
-      Name VARCHAR(100) NOT NULL,
-      Surname VARCHAR(100) NOT NULL,
-      Nickname VARCHAR(100) DEFAULT NULL,
-      Invitees INT(3) NOT NULL,
-      Cell VARCHAR(12) DEFAULT NULL,
-      Email VARCHAR(100) DEFAULT NULL,
-      Groups VARCHAR(100) DEFAULT NULL,
-      RSVP INT(3) DEFAULT NULL,
-      Ride BOOLEAN DEFAULT FALSE
-      ) DEFAULT CHARACTER SET utf8";
-
-      if (!mysqli_query($link, $sql)) {
-      $output = 'Error deleting user from Users table: ' . mysqli_error($link);
-      include 'output.html.php';
-      exit();
-      }
-      }
-
-      private function createMessages(int $eventID) {
-      $sql = "CREATE TABLE IF NOT EXISTS Messages_'$eventID' (
-      ID INT(2) NOT NULL AUTO_INCREMENT PRIMARY KEY,
-      MessageType VARCHAR(100) NOT NULL,
-      Message TEXT NOT NULL,
-      Groups VARCHAR(100) DEFAULT NULL,
-      SendDate DATE NOT NULL,
-      SendTime TIME NOT NULL
-      ) DEFAULT CHARACTER SET utf8";
-
-      if (!mysqli_query($link, $sql)) {
-      $output = 'Error deleting user from Users table: ' . mysqli_error($link);
-      include 'output.html.php';
-      exit();
-      }
-      }
-
-      private function createRawData(int $eventID) {
-      $sql = "CREATE TABLE IF NOT EXISTS RawData_'$eventID' (
-      ID INT(10) NOT NULL AUTO_INCREMENT PRIMARY KEY,
-      Name VARCHAR(100) NOT NULL,
-      Surname VARCHAR(100) NOT NULL,
-      Cell VARCHAR(12) DEFAULT NULL,
-      Email VARCHAR(100) DEFAULT NULL,
-      Groups VARCHAR(100) DEFAULT NULL,
-      RSVP INT(3) DEFAULT NULL,
-      Message TEXT NOT NULL,
-      RecivedDate DATE NOT NULL,
-      RecivedTime TIME NOT NULL
-      ) DEFAULT CHARACTER SET utf8";
-
-      if (!mysqli_query($link, $sql)) {
-      $output = 'Error deleting user from Users table: ' . mysqli_error($link);
-      include 'output.html.php';
-      exit();
-      }
-      }
-     */
-
     /**
      * __construct: create new Event object. if Event not in Events table (eventName,eventDate!=Null) add Event to Events table (do not make any change to Users Table!)
      * @param User $user : user element connected to this instance of event
      * @param string $EventName : name of event owner/owners or name of event
      * @param date $EventDate : date of event
-     * @param string $Email : Email to use for sending and receiving Emails
-     * @param string $EventPhone : Phone to use for sending and receiving messages
+     * @param string $EventPhone : Phone to use for sending and receiving messages 
+     * @param string $EventEmail : Email to use for sending and receiving Emails
      * @return object Event
      */
-    public function __construct(User $user, $EventName = NULL, $EventDate = NULL, $EventPhone = NULL, $EventEmail = NULL) {
+    public function __construct(User $user, $EventID = NULL, $EventName = NULL, $EventDate = NULL, $EventPhone = NULL, $EventEmail = NULL) {
 
         if (!isset(self::$db)) {
             self::$db = $user->getDB();
         }
+        
+        // user exists          ??? when using new user($ID)
+        if ($events = $user->getEvents() and $EventID === NULL and $events['event1'] != NULL) {
+            $this->eventID = $events['event1'];
+        }
+
+        if ($EventID != NULL and ! isset($this->eventID)) {
+            $this->eventID = $EventID;
+        }
+
         // Event is not in Events table (new Event)
-        if ($EventName != NULL and $EventDate != NULL and $EventDate != NULL and $EventEmail != NULL and $EventPhone != NULL) {
+        elseif ($EventName != NULL and $EventDate != NULL and $EventEmail != NULL and $EventPhone != NULL) {
             // initiate Database with user Database
             // Make strings query safe
             $rootID = $user->getID();
@@ -111,12 +64,15 @@ class Event implements iEvent {
             if (!isset($this->eventID)) {
                 $this->eventID = self::$db->insertID();
             }
+            
             // make new RSVP, Messages and RawData tables
+            $this->rsvp = new RSVP($this);
             /*
-              $rsvp = new RSVP();
-              $messages = new Messages();
-              $rawData = new RawData();
+              $this->messages = new Messages();
+              $this->rawData = new RawData();
              */
+        } else {
+            //throw exeption
         }
         // Event is in Events table
         return;
@@ -131,28 +87,48 @@ class Event implements iEvent {
         // Check user permission for event
         $result = $user->getEvents();
         for ($i = 1; $i <= 3; $i++) {
-            if ($result["event$i"] === $this->eventID) {
-                $eventID = $this->eventID;
-                if ($result["permission$i"] === 'root') {
+            $eventID = $this->eventID;
+            if ($result["event$i"] === $this->eventID and $result["permission$i"] === 'root') {
+                
                     // delete event from Events table
                     $sql = self::$db->query("DELETE FROM Events WHERE ID=$eventID");
-                    if ($sql) {
-                        /*
-                          // delete RSVP[eventID] table
-                          $sql = $db->query("DROP TABLE RSVP$eventID");
-                          // delete Messages[eventID] table
-                          $sql = $db->query("DROP TABLE Messages$eventID");
-                          // delete RawData[eventID] table
-                          $sql = $db->query("DROP TABLE RawData$eventID");
-                          if ($sql) {
-                         */
-                        self::$db->query("UPDATE Users SET Event$i=NULL, Permission$i=NULL
-                                            WHERE Event$i=$eventID");
-                        return true;
-                        //}
-                    }
-                }
+                    /*
+                      // delete RSVP[eventID] table
+                      $sqlRSVP = $this->rsvp->delete();
+                      // delete Messages[eventID] table
+                      $sqlMessages = $this->messages->delete();
+                      // delete RawData[eventID] table
+                      $sqlRawData = $this->rawData->delete();
+                      if ($sqlRSVP and $sqlMessages and $sqlRawData) {
+                     */
+                    
+                    //}
             }
+            $sql = self::$db->query("UPDATE Users SET Event$i=NULL, Permission$i=NULL
+                                            WHERE Event$i=$eventID");
+            // event deleted for user, shift all events left                    ??????
+        }
+        return true;
+    }
+
+    /**
+     * getDB:  get the DataBase
+     * @return type db (DataBase) / false if Database yet initialized
+     */
+    public function getDB() {
+        if (isset(self::$db)) {
+            return self::$db;
+        }
+        return false;
+    }
+
+    /**
+     * getEventID:  get Event ID
+     * @return int  Event ID / false if ID yet initialized
+     */
+    public function getEventID() {
+        if (isset(self::$eventID)) {
+            return self::$eventID;
         }
         return false;
     }
